Fix the issue with bitcoin stuck syncing on 100% (#375)
There were peers that are socket connected but do not respond with any message. The timer doesn't get started, since it starts only when it receives a message from the remote peer. That is why this type of peers never get timeouted. Starting timer for the first time when we send version message fix it. In a result, these type of peers will be disconnected.