Do not update state to syncing progress if there is no connection
There may be cases when state is first set to NotSynced and then to
Syncing when internet is lost. The reason is that blockSyncer may have
some blocks in queue to process when internet is lost. In this case we
need to set state to NotSynced. So we decided to set state to Syncing
only if internet is connected.