Update logic for selecting UTXO when there is a failed transaction
There may be a case when the failed transaction may become success after a while. For ex, the tx fails with low fee error, but then some miner decides to include it into the block despite of low fee. Since the tx gets failed status user may decide to re-send it. So if the kit selects other UTXOs in the second attempt, there will be 2 transactions. It may result in sending coins twice.

That is why when selecting UTXOs we need to use the failed ones first. It prevents from sending coins twice.

 - UnspentOutputSelectorSingleNoChange: Do not select output if there is at least one failed output
 - UnspentOutputSelector: First select the failed UTXOs